Fixes #15257 - tests moved from unit to functional
